requirements.txt
gradio>=4.0.0
pyyaml>=6.0
pydantic>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
anthropic>=0.25.0
openai>=1.0.0
pytest>=7.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
httpx>=0.27.0
